backend/requirements.txt
# Nexo Sinérgico Backend Dependencies
# Python 3.9+

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0  # Async PostgreSQL driver

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# AI APIs
google-generativeai>=0.8.3

# Cloud Storage
cloudinary==1.36.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Monitoring & Observability
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.0.0

passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
